# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Believer Backend API
  version: 1.8.0
  description: Active REST API contract for auth, account recovery, account-governance settings actions, business listings, mosques, backend-owned mosque prayer times, mosque page content, mosque image uploads, bookmarks, notifications, and services.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request payload
        details:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          type: object
          example: {}
    PaginationMeta:
      type: object
      required: [pagination]
      properties:
        pagination:
          type: object
          required: [page, limit, total, hasNext]
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 42
            hasNext:
              type: boolean
              example: true
...
    SuccessFlagEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [success]
          properties:
            success:
              type: boolean
              example: true
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
...
    BookmarkCreateRequest:
      type: object
      required: [mosqueId]
      properties:
        mosqueId:
          type: string
          format: uuid
    BookmarkCreatePayload:
      type: object
      required: [mosqueId, status]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        mosqueId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [created, already_bookmarked]
    BookmarkCreateEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/BookmarkCreatePayload'
        error:
          nullable: true
          example: null
        meta:
          type: object
          example: {}
    BookmarkListItem:
      type: object
      required:
        [id, name, city, country, latitude, longitude, facilities, isVerified, bookmarkedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        addressLine:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
          nullable: true
        country:
          type: string
        postalCode:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        facilities:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        bookmarkedAt:
          type: string
          format: date-time
    BookmarkListPayload:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookmarkListItem'
    BookmarkListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: '#/components/schemas/BookmarkListPayload'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
paths:
...
  /api/v1/bookmarks:
    get:
      tags: [Bookmarks]
      summary: List user bookmarks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Bookmark list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkListEnvelope'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Bookmarks]
      summary: Save mosque bookmark
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkCreateRequest'
      responses:
        '200':
          description: Already bookmarked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkCreateEnvelope'
        '201':
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkCreateEnvelope'
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Mosque not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/bookmarks/{mosqueId}:
    delete:
      tags: [Bookmarks]
      summary: Remove bookmark
      security:
        - bearerAuth: []
      parameters:
        - name: mosqueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bookmark removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessFlagEnvelope'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...